move (set|get)(translate|scale|pointgroupscale) into components/drawing
