+++ test/jasmine/tests/lib_test.js
@@ -1243,210 +1243,6 @@ describe('Test lib.js:', function() {
         });
     });
 
-    describe('getTranslate', function() {
-
-        it('should work with regular DOM elements', function() {
-            var el = document.createElement('div');
-
-            expect(Lib.getTranslate(el)).toEqual({ x: 0, y: 0 });
-
-            el.setAttribute('transform', 'translate(123.45px, 67)');
-            expect(Lib.getTranslate(el)).toEqual({ x: 123.45, y: 67 });
-
-            el.setAttribute('transform', 'translate(123.45)');
-            expect(Lib.getTranslate(el)).toEqual({ x: 123.45, y: 0 });
-
-            el.setAttribute('transform', 'translate(1 2)');
-            expect(Lib.getTranslate(el)).toEqual({ x: 1, y: 2 });
-
-            el.setAttribute('transform', 'translate(1 2); rotate(20deg)');
-            expect(Lib.getTranslate(el)).toEqual({ x: 1, y: 2 });
-
-            el.setAttribute('transform', 'rotate(20deg) translate(1 2);');
-            expect(Lib.getTranslate(el)).toEqual({ x: 1, y: 2 });
-
-            el.setAttribute('transform', 'rotate(20deg)');
-            expect(Lib.getTranslate(el)).toEqual({ x: 0, y: 0 });
-        });
-
-        it('should work with d3 elements', function() {
-            var el = d3.select(document.createElement('div'));
-
-            el.attr('transform', 'translate(123.45px, 67)');
-            expect(Lib.getTranslate(el)).toEqual({ x: 123.45, y: 67 });
-
-            el.attr('transform', 'translate(123.45)');
-            expect(Lib.getTranslate(el)).toEqual({ x: 123.45, y: 0 });
-
-            el.attr('transform', 'translate(1 2)');
-            expect(Lib.getTranslate(el)).toEqual({ x: 1, y: 2 });
-
-            el.attr('transform', 'translate(1 2); rotate(20)');
-            expect(Lib.getTranslate(el)).toEqual({ x: 1, y: 2 });
-
-            el.attr('transform', 'rotate(20)');
-            expect(Lib.getTranslate(el)).toEqual({ x: 0, y: 0 });
-        });
-    });
-
-    describe('setTranslate', function() {
-
-        it('should work with regular DOM elements', function() {
-            var el = document.createElement('div');
-
-            Lib.setTranslate(el, 5);
-            expect(el.getAttribute('transform')).toBe('translate(5, 0)');
-
-            Lib.setTranslate(el, 10, 20);
-            expect(el.getAttribute('transform')).toBe('translate(10, 20)');
-
-            Lib.setTranslate(el);
-            expect(el.getAttribute('transform')).toBe('translate(0, 0)');
-
-            el.setAttribute('transform', 'translate(0, 0); rotate(30)');
-            Lib.setTranslate(el, 30, 40);
-            expect(el.getAttribute('transform')).toBe('rotate(30) translate(30, 40)');
-        });
-
-        it('should work with d3 elements', function() {
-            var el = d3.select(document.createElement('div'));
-
-            Lib.setTranslate(el, 5);
-            expect(el.attr('transform')).toBe('translate(5, 0)');
-
-            Lib.setTranslate(el, 30, 40);
-            expect(el.attr('transform')).toBe('translate(30, 40)');
-
-            Lib.setTranslate(el);
-            expect(el.attr('transform')).toBe('translate(0, 0)');
-
-            el.attr('transform', 'translate(0, 0); rotate(30)');
-            Lib.setTranslate(el, 30, 40);
-            expect(el.attr('transform')).toBe('rotate(30) translate(30, 40)');
-        });
-    });
-
-    describe('getScale', function() {
-
-        it('should work with regular DOM elements', function() {
-            var el = document.createElement('div');
-
-            expect(Lib.getScale(el)).toEqual({ x: 1, y: 1 });
-
-            el.setAttribute('transform', 'scale(1.23, 45)');
-            expect(Lib.getScale(el)).toEqual({ x: 1.23, y: 45 });
-
-            el.setAttribute('transform', 'scale(123.45)');
-            expect(Lib.getScale(el)).toEqual({ x: 123.45, y: 1 });
-
-            el.setAttribute('transform', 'scale(0.1 2)');
-            expect(Lib.getScale(el)).toEqual({ x: 0.1, y: 2 });
-
-            el.setAttribute('transform', 'scale(0.1 2); rotate(20deg)');
-            expect(Lib.getScale(el)).toEqual({ x: 0.1, y: 2 });
-
-            el.setAttribute('transform', 'rotate(20deg) scale(0.1 2);');
-            expect(Lib.getScale(el)).toEqual({ x: 0.1, y: 2 });
-
-            el.setAttribute('transform', 'rotate(20deg)');
-            expect(Lib.getScale(el)).toEqual({ x: 1, y: 1 });
-        });
-
-        it('should work with d3 elements', function() {
-            var el = d3.select(document.createElement('div'));
-
-            el.attr('transform', 'scale(1.23, 45)');
-            expect(Lib.getScale(el)).toEqual({ x: 1.23, y: 45 });
-
-            el.attr('transform', 'scale(123.45)');
-            expect(Lib.getScale(el)).toEqual({ x: 123.45, y: 1 });
-
-            el.attr('transform', 'scale(0.1 2)');
-            expect(Lib.getScale(el)).toEqual({ x: 0.1, y: 2 });
-
-            el.attr('transform', 'scale(0.1 2); rotate(20)');
-            expect(Lib.getScale(el)).toEqual({ x: 0.1, y: 2 });
-
-            el.attr('transform', 'rotate(20)');
-            expect(Lib.getScale(el)).toEqual({ x: 1, y: 1 });
-        });
-    });
-
-    describe('setScale', function() {
-
-        it('should work with regular DOM elements', function() {
-            var el = document.createElement('div');
-
-            Lib.setScale(el, 5);
-            expect(el.getAttribute('transform')).toBe('scale(5, 1)');
-
-            Lib.setScale(el, 30, 40);
-            expect(el.getAttribute('transform')).toBe('scale(30, 40)');
-
-            Lib.setScale(el);
-            expect(el.getAttribute('transform')).toBe('scale(1, 1)');
-
-            el.setAttribute('transform', 'scale(1, 1); rotate(30)');
-            Lib.setScale(el, 30, 40);
-            expect(el.getAttribute('transform')).toBe('rotate(30) scale(30, 40)');
-        });
-
-        it('should work with d3 elements', function() {
-            var el = d3.select(document.createElement('div'));
-
-            Lib.setScale(el, 5);
-            expect(el.attr('transform')).toBe('scale(5, 1)');
-
-            Lib.setScale(el, 30, 40);
-            expect(el.attr('transform')).toBe('scale(30, 40)');
-
-            Lib.setScale(el);
-            expect(el.attr('transform')).toBe('scale(1, 1)');
-
-            el.attr('transform', 'scale(0, 0); rotate(30)');
-            Lib.setScale(el, 30, 40);
-            expect(el.attr('transform')).toBe('rotate(30) scale(30, 40)');
-        });
-    });
-
-    describe('setPointGroupScale', function() {
-        var el, sel;
-
-        beforeEach(function() {
-            el = document.createElement('div');
-            sel = d3.select(el);
-        });
-
-        it('sets the scale of a point', function() {
-            Lib.setPointGroupScale(sel, 2, 2);
-            expect(el.getAttribute('transform')).toBe('scale(2,2)');
-        });
-
-        it('appends the scale of a point', function() {
-            el.setAttribute('transform', 'translate(1,2)');
-            Lib.setPointGroupScale(sel, 2, 2);
-            expect(el.getAttribute('transform')).toBe('translate(1,2) scale(2,2)');
-        });
-
-        it('modifies the scale of a point', function() {
-            el.setAttribute('transform', 'translate(1,2) scale(3,4)');
-            Lib.setPointGroupScale(sel, 2, 2);
-            expect(el.getAttribute('transform')).toBe('translate(1,2) scale(2,2)');
-        });
-
-        it('does not apply the scale of a point if scale (1, 1)', function() {
-            el.setAttribute('transform', 'translate(1,2)');
-            Lib.setPointGroupScale(sel, 1, 1);
-            expect(el.getAttribute('transform')).toBe('translate(1,2)');
-        });
-
-        it('removes the scale of a point if scale (1, 1)', function() {
-            el.setAttribute('transform', 'translate(1,2) scale(3,4)');
-            Lib.setPointGroupScale(sel, 1, 1);
-            expect(el.getAttribute('transform')).toBe('translate(1,2)');
-        });
-    });
-
     describe('pushUnique', function() {
 
         beforeEach(function() {
